Add SerializedMetadata column to ActivityExecutionRecords
This commit introduces a new column named `SerializedMetadata` to the `ActivityExecutionRecords` table. A migration class `AddSerializedMetadataToActivityExecutionRecords` has been created, which includes methods to add and remove the column. The `Up` method adds the column with a type of `nvarchar(max)` and allows null values, while the `Down` method removes it if the migration is rolled back. The designer file has also been updated to reflect this change in the model builder configuration.